Updating blue-green-deployment Demo for v0.6
While I was running through this demo on GKE with knative 0.6 I ran into a problem where I couldn't apply the configurations as the api `serving.knative.dev/v1alpha1` was not what I had installed via the [Installing knative on GKE v0.6]() documentation. 

Then after that I wasn't able to access the new revision at the url as later version of knative deprecated the `name` field on traffic definitions in favor of the `tag` field.

This change updates both of those things accordingly to bring the documentation up to date with v0.6 apis.